Fixes for collection determination
The mechanism for determining whether the object is a collection or a singular
item is not ideal.

Use case:

Produce a representation for a tree structure in which the attributes for each
node are determined on whether the node has any children or not, and including
a representation of the nodes children is determined by attributes on the
node. In this case, precise control over when to iterate into the children is
required.