Don't convert collections to an array
There are other types of collections than Arrays. DataMapper::Collection is
one example of this.

There are differences between between iterating over a DataMapper::Collection
and converting it to an Array and iterating over that Array.

One of the differences is in which Collection each resource believes it
belongs to. When iterating over a Collection, each resource's Collection is
the collection being iterated over. However, if the Collection is converted to
an Array, each resource's Collection is a new Collection containing a single
resource.

This interferes with lazy loading, resulting in O(n) trips to load the
attributes rather than a single trip.

To fix this, let's trust the developer to provide a collection when they use
the 'collection' method.